Big progress on testing, slightly changed the directory structure so that it no longer includes the path 'unit_tests' as I'm also doing integrating testing (I think). At this point I have all the coverage I want on the CUDA code. Moving on to C.
