--- cmake_testing/unit_tests/CMakeLists.txt
@@ -1,3 +1,0 @@
-# add_subdirectory(fundamental_coords)
-# add_subdirectory(primary_beam_cuda)
-add_subdirectory(source_components)
